Change stats rate types from Float64 to UInt32 for consistency
Updates rate variables and logs to use UInt32 instead of Float64 in both
main and nodes controllers to maintain consistency with the 1s interval change.

+++ src/lavinmq/http/controller/main.cr
@@ -19,16 +19,16 @@ module LavinMQ
           recv_rate, send_rate = 0_f64, 0_f64
           ready_log = Deque(UInt32).new(LavinMQ::Config.instance.stats_retention_seconds)
           unacked_log = Deque(UInt32).new(LavinMQ::Config.instance.stats_retention_seconds)
-          recv_rate_log = Deque(Float64).new(LavinMQ::Config.instance.stats_retention_seconds)
-          send_rate_log = Deque(Float64).new(LavinMQ::Config.instance.stats_retention_seconds)
+          recv_rate_log = Deque(UInt32).new(LavinMQ::Config.instance.stats_retention_seconds)
+          send_rate_log = Deque(UInt32).new(LavinMQ::Config.instance.stats_retention_seconds)
           {% for name in OVERVIEW_STATS %}
           {{name.id}}_count = 0_u64
-          {{name.id}}_rate = 0_f64
-          {{name.id}}_log = Deque(Float64).new(LavinMQ::Config.instance.stats_retention_seconds)
+          {{name.id}}_rate = 0_u32
+          {{name.id}}_log = Deque(UInt32).new(LavinMQ::Config.instance.stats_retention_seconds)
           {% end %}
           {% for name in CHURN_STATS %}
           {{name.id}} = 0_u64
-          {{name.id}}_rate = 0_f64
+          {{name.id}}_rate = 0_u32
           {% end %}
 
           vhosts(user(context)).each do |vhost|
